create new stats query service for valkyrie that defaults to original query service for wings/fedora
